refactor: AverageAgeScore enum 클래스 외부로 이동 - Result 도메인에서 분리
