+++ src/main/java/com/b4after/bntquestion/domain/Result.java
@@ -1,8 +1,6 @@
 package com.b4after.bntquestion.domain;
 
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -28,25 +26,7 @@ public class Result {
                 )
                 .collect(Collectors.toList());
     }
-    @Getter
-    @RequiredArgsConstructor
-    private enum AgeAverageScore {
-        AGED_GROUP1(12.37, 55, 64),
-        AGED_GROUP2(11.17, 65, 74),
-        AGED_GROUP3(10.50, 75, 84),
-        AGED_GROUP4(6.64, 85, Integer.MAX_VALUE);
-        private final double average;
-        private final int minAge;
-        private final int maxAge;
 
-        public static double getAverageScore(int age) {
-            AgeAverageScore ageAverage = Arrays.stream(AgeAverageScore.values())
-                    .filter(range -> age >= range.minAge && age <= range.maxAge)
-                    .findFirst()
-                    .orElseThrow(() -> new IllegalArgumentException("범위 밖입니다."));
-            return ageAverage.getAverage();
-            }
-        }
     private int calculateScore(List<Answer> answers) {
         return (int) answers.stream()
                 .filter(Answer::isCorrect)
